+++ java-manta-client/src/main/java/com/joyent/manta/http/RequestIdInterceptor.java
@@ -7,6 +7,9 @@
  */
 package com.joyent.manta.http;
 
+import com.fasterxml.uuid.EthernetAddress;
+import com.fasterxml.uuid.Generators;
+import com.fasterxml.uuid.impl.TimeBasedGenerator;
 import org.apache.http.Header;
 import org.apache.http.HttpException;
 import org.apache.http.HttpRequest;
@@ -16,6 +19,10 @@ import org.apache.http.protocol.HttpContext;
 import org.slf4j.MDC;
 
 import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
+import java.security.NoSuchProviderException;
+import java.security.SecureRandom;
+import java.util.Random;
 import java.util.UUID;
 
 /**
@@ -28,13 +35,34 @@ import java.util.UUID;
  */
 public class RequestIdInterceptor implements HttpRequestInterceptor {
     /**
+     * Time-based UUID generator for generating request ids.
+     */
+    private static final TimeBasedGenerator TIME_BASED_GENERATOR;
+
+    static {
+        Random nonBlockingRandomness;
+
+        try {
+            nonBlockingRandomness = SecureRandom.getInstance("NativePRNGNonBlocking", "SUN");
+        } catch (NoSuchAlgorithmException | NoSuchProviderException e) {
+            nonBlockingRandomness = new Random(System.nanoTime());
+        }
+
+        // Fake ethernet address based on a random value
+        final EthernetAddress ethernetAddress = EthernetAddress.constructMulticastAddress(
+                nonBlockingRandomness);
+        TIME_BASED_GENERATOR = Generators.timeBasedGenerator(ethernetAddress);
+    }
+
+    /**
      * Constant identifying the request id as a MDC attribute.
      */
     public static final String MDC_REQUEST_ID_STRING = "mantaRequestId";
 
     @Override
     public void process(final HttpRequest request, final HttpContext context) throws HttpException, IOException {
-        final String requestId = UUID.randomUUID().toString();
+        final UUID id = TIME_BASED_GENERATOR.generate();
+        final String requestId = id.toString();
         final Header idHeader = new BasicHeader(MantaHttpHeaders.REQUEST_ID, requestId);
         request.addHeader(idHeader);